+++ application/Hardware/Axes/XAxis/XAxis.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IXAxis.hpp"
-#include "application/Logger/ILogger.hpp"
+#include "application/System/Logger/ILogger.hpp"
 
 namespace ATC {
 class XAxis : public IXAxis {
