+++ application/Hardware/Axes/YAxis/YAxis.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "IYAxis.hpp"
-#include "application/Logger/ILogger.hpp"
+#include "application/System/Logger/ILogger.hpp"
 
 #include <functional>
 #include <stdint.h>
