+++ application/Hardware/Rotator/DcMotorRotator/DcMotorRotator.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "../IRotator.hpp"
-#include "application/Logger/ILogger.hpp"
+#include "application/System/Logger/ILogger.hpp"
 
 namespace ATC {
 class DcMotorRotator : public IRotator {
