+++ application/System/Logger/LogFormatter.hpp
