+++ application/System/Logger/UartLogger/UartLogger.cpp
