+++ application/System/Logger/UartLogger/UartLogger.hpp
