+++ atc-nucleo-u5a5zj/Core/Inc/TargetBoard.hpp
@@ -4,8 +4,6 @@
 #include "Adapters/Spi/Spi.hpp"
 #include "Adapters/SystemClock/SystemClock.hpp"
 #include "Adapters/Uart/Uart.hpp"
-#include "application/Board/Board.hpp"
-#include "application/Board/BoardDevices.hpp"
 #include "application/Hardware/Axes/XAxis/XAxis.hpp"
 #include "application/Hardware/Axes/YAxis/YAxis.hpp"
 #include "application/Hardware/Axes/ZAxis/ZAxis.hpp"
@@ -14,7 +12,9 @@
 #include "application/Hardware/Heater/ThermistorHeater/ThermistorHeater.hpp"
 #include "application/Hardware/Rotator/DcMotorRotator/DcMotorRotator.hpp"
 #include "application/Hardware/TouchController/Xpt2046TouchController/Xpt2046TouchController.hpp"
-#include "application/Logger/UartLogger/UartLogger.hpp"
+#include "application/System/Board/Board.hpp"
+#include "application/System/Board/BoardDevices.hpp"
+#include "application/System/Logger/UartLogger/UartLogger.hpp"
 #include "main.h"
 
 extern SPI_HandleTypeDef hspi1;
